Implementing an Actor Model within SuspendedByteChannels
SuspendedByteChannels will require some form of synchronization to better maintain their mutable state. This is obvious in the case of the SecureSocketChannel as it will require a synchronized usage of it's four ByteBuffers. These ByteBuffers cannot properly mutate with multiple operations happening at the same time. An Actor design will help funnel operations in a single pipeline.